+++ AuthService/cookieUser.txt
@@ -1,0 +1,5 @@
+# Netscape HTTP Cookie File
+# https://curl.haxx.se/docs/http-cookies.html
+# This file was generated by libcurl! Edit at your own risk.
+
+#HttpOnly_localhost	FALSE	/	FALSE	0	JSESSIONID	11C91DD5BC09B0E9FB9055CD5274C0C7
