+++ AuthService/src/main/kotlin/dev/easypass/auth/AuthApplication.kt
@@ -1,13 +1,12 @@
 package dev.easypass.auth
 
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Primary
-import org.springframework.util.ResourceUtils
-import java.io.FileInputStream
+import org.springframework.boot.*
+import org.springframework.boot.autoconfigure.*
+import org.springframework.cloud.client.discovery.*
+import org.springframework.cloud.netflix.zuul.*
+import org.springframework.context.annotation.*
+import org.springframework.util.*
+import java.io.*
 import java.util.*
 
 @SpringBootApplication
