+++ AuthService/src/main/kotlin/dev/easypass/auth/security/challenge/InternalChallenge.kt
@@ -1,8 +1,7 @@
 package dev.easypass.auth.security.challenge
 
-import dev.easypass.auth.security.EncryptionLibrary
-import java.time.Duration
-import java.time.LocalDateTime
+import dev.easypass.auth.security.*
+import java.time.*
 import java.util.*
 
 /**
