+++ AuthService/src/main/kotlin/dev/easypass/auth/security/exception/NoActiveChallengeException.kt
@@ -1,6 +1,6 @@
 package dev.easypass.auth.security.exception
 
-import org.springframework.security.core.AuthenticationException
+import org.springframework.security.core.*
 
 /**
  * This Exception is thrown if the unique identifier of an object is already saved in the database
