+++ AuthService/src/main/kotlin/dev/easypass/auth/security/handler/RestAuthenticationEntryPoint.kt
@@ -1,10 +1,9 @@
 package dev.easypass.auth.security.handler
 
-import org.springframework.security.core.AuthenticationException
-import org.springframework.security.web.AuthenticationEntryPoint
-import java.io.IOException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import org.springframework.security.core.*
+import org.springframework.security.web.*
+import java.io.*
+import javax.servlet.http.*
 
 
 class RestAuthenticationEntryPoint : AuthenticationEntryPoint {
